+++ config/vim/pack/plugins/start/fzf/src/options.go
@@ -56,7 +56,7 @@ Usage: fzf [options]
     --sync                   Synchronous search for multi-staged filtering
 
   GLOBAL STYLE
-    --style=PRESET           Apply a style preset [default|minimal|full]
+    --style=PRESET           Apply a style preset [default|minimal|full[:BORDER_STYLE]
     --color=COLSPEC          Base scheme (dark|light|16|bw) and/or custom colors
     --no-color               Disable colors
     --no-bold                Do not use bold text
@@ -97,6 +97,8 @@ Usage: fzf [options]
     --track                  Track the current selection when the result is updated
     --tac                    Reverse the order of the input
     --gap[=N]                Render empty lines between each item
+    --gap-line[=STR]         Draw horizontal line on each gap using the string
+                             (default: '┈' or '-')
     --keep-right             Keep the right end of the line visible on overflow
     --scroll-off=LINES       Number of screen lines to keep above or below when
                              scrolling to the top or to the bottom (default: 0)
@@ -127,7 +129,8 @@ Usage: fzf [options]
     --info=STYLE             Finder info style
                              [default|right|hidden|inline[-right][:PREFIX]]
     --info-command=COMMAND   Command to generate info line
-    --separator=STR          String to form horizontal separator on info line
+    --separator=STR          Draw horizontal separator on info line using the string
+                             (default: '─' or '-')
     --no-separator           Hide info line separator
     --filepath-word          Make word-wise movements respect path separators
     --input-border[=STYLE]   Draw border around the input section
@@ -210,6 +213,9 @@ Usage: fzf [options]
     FZF_API_KEY              X-API-Key header for HTTP server (--listen)
 
 `
+
+// Can be changed by --style
+var defaultBorderShape tui.BorderShape = tui.DefaultBorderShape
 
 const defaultInfoPrefix = " < "
 
@@ -582,6 +588,7 @@ type Options struct {
 	HeaderLines       int
 	HeaderFirst       bool
 	Gap               int
+	GapLine           *string
 	Ellipsis          *string
 	Scrollbar         *string
 	Margin            [4]sizeSpec
@@ -624,7 +631,7 @@ func filterNonEmpty(input []string) []string {
 }
 
 func defaultPreviewOpts(command string) previewOpts {
-	return previewOpts{command, posRight, sizeSpec{50, true}, "", false, false, false, false, true, tui.DefaultBorderShape, 0, 0, nil}
+	return previewOpts{command, posRight, sizeSpec{50, true}, "", false, false, false, false, true, defaultBorderShape, 0, 0, nil}
 }
 
 func defaultOptions() *Options {
@@ -838,7 +845,7 @@ func parseBorder(str string, optional bool, allowLine bool) (tui.BorderShape, er
 		return tui.BorderNone, nil
 	}
 	if optional && str == "" {
-		return tui.DefaultBorderShape, nil
+		return defaultBorderShape, nil
 	}
 	return tui.BorderNone, errors.New("invalid border style (expected: rounded|sharp|bold|block|thinblock|double|horizontal|vertical|top|bottom|left|right|none)")
 }
@@ -1203,6 +1210,8 @@ func parseTheme(defaultTheme *tui.ColorTheme, str string) (*tui.ColorTheme, erro
 				mergeAttr(&theme.SelectedFg)
 			case "selected-bg":
 				mergeAttr(&theme.SelectedBg)
+			case "nth":
+				mergeAttr(&theme.Nth)
 			case "gutter":
 				mergeAttr(&theme.Gutter)
 			case "hl":
@@ -1253,6 +1262,8 @@ func parseTheme(defaultTheme *tui.ColorTheme, str string) (*tui.ColorTheme, erro
 				mergeAttr(&theme.Header)
 			case "header-bg":
 				mergeAttr(&theme.HeaderBg)
+			case "gap-line":
+				mergeAttr(&theme.GapLine)
 			default:
 				fail()
 			}
@@ -2567,6 +2578,15 @@ func parseOptions(index *int, opts *Options, allArgs []string) error {
 			}
 		case "--no-gap":
 			opts.Gap = 0
+		case "--gap-line":
+			if given, bar := optionalNextString(); given {
+				opts.GapLine = &bar
+			} else {
+				opts.GapLine = nil
+			}
+		case "--no-gap-line":
+			empty := ""
+			opts.GapLine = &empty
 		case "--ellipsis":
 			str, err := nextString("ellipsis string required")
 			if err != nil {
@@ -2714,7 +2734,7 @@ func parseOptions(index *int, opts *Options, allArgs []string) error {
 				return err
 			}
 		case "--style":
-			preset, err := nextString("preset name required: [default|minimal|full]")
+			preset, err := nextString("preset name required: [default|minimal|full[:BORDER_STYLE]]")
 			if err != nil {
 				return err
 			}
@@ -2867,12 +2887,15 @@ func parseOptions(index *int, opts *Options, allArgs []string) error {
 }
 
 func applyPreset(opts *Options, preset string) error {
+	// Reset to the platform default
+	defaultBorderShape = tui.DefaultBorderShape
+
 	switch strings.ToLower(preset) {
 	case "default":
 		opts.ListBorderShape = tui.BorderUndefined
 		opts.InputBorderShape = tui.BorderUndefined
 		opts.HeaderBorderShape = tui.BorderUndefined
-		opts.Preview.border = tui.DefaultBorderShape
+		opts.Preview.border = defaultBorderShape
 		opts.Preview.info = true
 		opts.InfoStyle = infoDefault
 		opts.Theme.Gutter = tui.NewColorAttr()
@@ -2891,19 +2914,29 @@ func applyPreset(opts *Options, preset string) error {
 		opts.Separator = &empty
 		opts.Scrollbar = &empty
 		opts.CursorLine = false
-	case "full":
-		opts.ListBorderShape = tui.DefaultBorderShape
-		opts.InputBorderShape = tui.DefaultBorderShape
-		opts.HeaderBorderShape = tui.DefaultBorderShape
-		opts.Preview.border = tui.DefaultBorderShape
+	default:
+		tokens := strings.SplitN(preset, ":", 2)
+		if tokens[0] != "full" {
+			return errors.New("unsupported style preset: " + preset)
+		}
+		if len(tokens) == 2 && len(tokens[1]) > 0 {
+			var err error
+			defaultBorderShape, err = parseBorder(tokens[1], false, false)
+			if err != nil {
+				return err
+			}
+		}
+
+		opts.ListBorderShape = defaultBorderShape
+		opts.InputBorderShape = defaultBorderShape
+		opts.HeaderBorderShape = defaultBorderShape
+		opts.Preview.border = defaultBorderShape
 		opts.Preview.info = true
 		opts.InfoStyle = infoInlineRight
 		opts.Theme.Gutter = tui.NewColorAttr()
 		opts.Separator = nil
 		opts.Scrollbar = nil
 		opts.CursorLine = true
-	default:
-		return errors.New("unsupported style preset: " + preset)
 	}
 	return nil
 }
@@ -2953,6 +2986,10 @@ func validateOptions(opts *Options) error {
 		}
 	}
 
+	if opts.Theme.Nth.IsColorDefined() {
+		return errors.New("only ANSI attributes are allowed for 'nth' (regular, bold, underline, reverse, dim, italic, strikethrough)")
+	}
+
 	return nil
 }
 
@@ -2985,6 +3022,14 @@ func postProcessOptions(opts *Options) error {
 			defaultPointer = ">"
 		}
 		opts.Pointer = &defaultPointer
+	}
+
+	if opts.GapLine == nil {
+		defaultGapLine := "┈"
+		if !opts.Unicode {
+			defaultGapLine = "-"
+		}
+		opts.GapLine = &defaultGapLine
 	}
 
 	markerLen := 1
@@ -3090,7 +3135,7 @@ func postProcessOptions(opts *Options) error {
 		boldify := func(c tui.ColorAttr) tui.ColorAttr {
 			dup := c
 			if (c.Attr & tui.AttrRegular) == 0 {
-				dup.Attr |= tui.Bold
+				dup.Attr |= tui.BoldForce
 			}
 			return dup
 		}